extensions/ProteinStructureFetch/vrprot/pypi_project/requirements.txt
pandas>=1.5.0
open3d>=0.15.1
Pillow>=9.2.0
requests>=2.28.1
trimesh>=3.15.3
